refactor(frontdesk): split poller.go by poll
poller.go held the Poller core and health polling alongside the three
other polls. Same package, pure moves with their doc comments:

- poller_announce.go: memberAnnounce, errAnnounceConflict,
  activeMemberCount, fleetPrimary, PollAnnounceOnce, announceToMember
- poller_traefik.go: PollTraefikOnce, noteTraefikAPIFailure,
  fetchTraefikServerStatus, traefikServiceInfo, parseTraefikServerStatus
- poller_versions.go: PollVersionsOnce, noteVersionFetchFailure,
  fetchMemberBuild, checkConfigStaleness, ConfigPollStale,
  ConfigPollWarm, checkAutoSyncStale

poller.go keeps the constants, the status types, Poller, NewPoller, the
setters, Run/tickLoop, health polling and the event helpers (930 -> 379
lines) and leaves the size allowlist. Tests are unchanged.
