+++ internal/frontdesk/poller.go
@@ -1,10 +1,7 @@
 package frontdesk
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"maps"
@@ -343,554 +340,6 @@ func (p *Poller) applyHealth(ctx context.Context, m *Member, hs HealthStatus) {
 			Type: "health.down", Severity: "error", Source: "frontdesk-poller",
 			Message: fmt.Sprintf("%s is unreachable after %d %s", m.Name, fails, util.Plural(fails, "check", "checks")), MemberID: m.ID,
 			Metadata: map[string]any{"error": hs.Error, "consecutive_failures": fails},
-		})
-	}
-}
-
-// memberAnnounce is the heartbeat body Front Desk POSTs to each member's
-// /api/fleet/announce. It carries only routing metadata: whether the member is
-// the fleet primary, and the primary's display name for the member's tooltip.
-type memberAnnounce struct {
-	IsPrimary   bool   `json:"is_primary"`
-	PrimaryName string `json:"primary_name,omitempty"`
-	// FrontdeskID is this Front Desk's persistent identity. A member records the
-	// first ID-bearing announcer as the owner of its fleet role and rejects
-	// announces from any other Front Desk while that owner is live. Empty on a
-	// legacy Front Desk build (the member then accepts unconditionally).
-	FrontdeskID string `json:"frontdesk_id,omitempty"`
-	// ActiveMembers is the fleet-wide count of StateActive members — the fair-share
-	// rate-limit divisor. Every active member is a Traefik round-robin backend, so
-	// each enforces 1/ActiveMembers of every configured limit. Omitted (0) by a
-	// legacy Front Desk build, which the member reads as divisor 1 (no division).
-	ActiveMembers int `json:"active_members,omitempty"`
-}
-
-// errAnnounceConflict is returned by announceToMember when a member rejects the
-// announce with 409: another Front Desk currently owns that member's fleet role.
-var errAnnounceConflict = errors.New("frontdesk: announce rejected (managed by another Front Desk)")
-
-// activeMemberCount counts members in StateActive — the exact set
-// BuildTraefikConfig puts behind the round-robin /v1 pool (traefik.go:96). It is
-// the fleet fair-share divisor each active member applies to its rate limits.
-// The divisor tracks the announced StateActive set, which is the same set
-// BuildTraefikConfig routes to — but Traefik's live pool can diverge from it
-// transiently: it may eject a StateActive backend its own health check finds
-// unhealthy, or pick up a state change on a different schedule than the ~5s
-// announce loop. That skew is bounded and self-correcting on the next
-// announce, and its dominant direction is safe: when Traefik routes to fewer
-// backends than N, each survivor divides by a too-large N and the fleet
-// under-serves (never exceeds the global cap). The opposite (brief over-cap)
-// only occurs on an Active->Drained transition where Traefik still routes to
-// the old member for a beat after N dropped — the accepted membership-change
-// blip, not a sustained violation.
-func activeMemberCount(members []*Member) int {
-	n := 0
-	for _, m := range members {
-		if m.State == StateActive {
-			n++
-		}
-	}
-	return n
-}
-
-// fleetPrimary resolves which member the fleet's primary is, for the announce,
-// and returns its current name from the live roster.
-//
-// Two sources can name a primary, and which one is authoritative depends on
-// whether auto-sync is running. While it is, the operator's designation is the
-// live answer: it is what the loop pushes from, and it takes effect the moment
-// they repoint, where the last-sync marker still names whichever member drove the
-// previous run. With auto-sync off the designation is dormant configuration, it
-// survives being switched off (clearing it needs a confirmed token), and the last
-// wizard run is then the more recent operator act, so the marker wins.
-//
-// A candidate that is not in the roster is skipped rather than returned: a
-// designation left pointing at a removed member would otherwise beat a marker that
-// still names a real one, and every member would be told there is no primary.
-//
-// Nothing resolving means no member is flagged primary. The membership signal is
-// still worth sending, so the caller continues without one rather than aborting; a
-// read error is treated the same way.
-func (p *Poller) fleetPrimary(ctx context.Context, members []*Member) (id, name string, ok bool) {
-	cfg, cfgErr := p.store.GetAutoSync(ctx)
-	if cfgErr != nil {
-		debuglog.Warn("frontdesk: poll announce: read auto-sync config", "error", cfgErr)
-	}
-	designated := ""
-	if cfgErr == nil {
-		designated = cfg.PrimaryID
-	}
-	state, hasMarker, stateErr := p.store.GetFleetSyncState(ctx)
-	if stateErr != nil {
-		debuglog.Warn("frontdesk: poll announce: fleet sync state", "error", stateErr)
-		hasMarker = false
-	}
-	marked := ""
-	if hasMarker {
-		marked = state.PrimaryID
-	}
-
-	// designated is empty unless the read succeeded, so the error case needs no arm
-	// of its own here: it simply contributes no candidate.
-	var candidates []string
-	if designated != "" && cfg.Enabled {
-		candidates = append(candidates, designated)
-	}
-	if marked != "" {
-		candidates = append(candidates, marked)
-	}
-	// A designation with auto-sync off still beats naming nobody.
-	if designated != "" && !cfg.Enabled {
-		candidates = append(candidates, designated)
-	}
-
-	for _, want := range candidates {
-		for _, m := range members {
-			if m.ID == want {
-				return m.ID, m.Name, true
-			}
-		}
-	}
-	return "", "", false
-}
-
-// PollAnnounceOnce tells every reachable, tokened member that Front Desk is in
-// contact, and which member is the fleet primary. It is the producing half of
-// HA Phase 6: a member uses these announces to light up the HA line on its own
-// dashboard and to self-clear it when they stop. Best-effort, exactly like the
-// health poll: a member that is down, has no stored token, or runs an older
-// build without the endpoint is silently skipped, never retried or surfaced.
-func (p *Poller) PollAnnounceOnce(ctx context.Context) {
-	members, err := p.store.ListMembers(ctx)
-	if err != nil {
-		debuglog.Warn("frontdesk: poll announce: list members", "error", err)
-		return
-	}
-	// Resolved against the live roster, so a renamed primary announces its current
-	// name and a designation pointing at a removed member never wins.
-	primaryID, primaryName, hasPrimary := p.fleetPrimary(ctx, members)
-	activeCount := activeMemberCount(members)
-	for _, m := range members {
-		token, ok, err := p.store.MemberToken(ctx, m.ID)
-		if err != nil || !ok {
-			continue // no stored token: the announce endpoint needs admin auth
-		}
-		ann := memberAnnounce{
-			IsPrimary:     hasPrimary && m.ID == primaryID,
-			PrimaryName:   primaryName,
-			FrontdeskID:   p.frontdeskID,
-			ActiveMembers: activeCount,
-		}
-		if err := p.announceToMember(ctx, m.URL, token, ann); err != nil {
-			if errors.Is(err, errAnnounceConflict) {
-				// Another Front Desk owns this member. Warn once per member per
-				// process run (announces fire every ~5s; without the latch this
-				// would be a log line every poll). The latch is never reset: the
-				// conflict is a persistent misconfiguration the operator resolves.
-				p.mu.Lock()
-				already := p.conflictNotified[m.ID]
-				if !already {
-					p.conflictNotified[m.ID] = true
-				}
-				p.mu.Unlock()
-				if !already {
-					debuglog.Warn("frontdesk: member is managed by another Front Desk; announce rejected", "member", m.ID)
-				}
-				continue
-			}
-			debuglog.Debug("frontdesk: announce to member failed", "member", m.ID, "error", err)
-		}
-	}
-}
-
-// announceToMember POSTs one heartbeat through the guarded probe client (the
-// same SSRF-protected client the health poll uses), carrying the member's admin
-// Bearer token. A non-204 reply is an error so the caller can log-and-continue.
-func (p *Poller) announceToMember(ctx context.Context, baseURL, token string, ann memberAnnounce) error {
-	body, err := json.Marshal(ann)
-	if err != nil {
-		return err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, baseURL+memberAnnouncePath, bytes.NewReader(body))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	req.Header.Set("Content-Type", "application/json")
-	resp, err := p.client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer func() { _ = resp.Body.Close() }()
-	_, _ = io.Copy(io.Discard, io.LimitReader(resp.Body, 1<<10))
-	if resp.StatusCode == http.StatusConflict {
-		// The member is owned by another Front Desk. Distinguish this from a
-		// generic failure so the caller can surface it once, not spam Debug.
-		return errAnnounceConflict
-	}
-	if resp.StatusCode != http.StatusNoContent {
-		return fmt.Errorf("announce returned %d", resp.StatusCode)
-	}
-	return nil
-}
-
-// PollTraefikOnce fetches Traefik's serverStatus and maps it onto members by URL.
-func (p *Poller) PollTraefikOnce(ctx context.Context) {
-	if p.traefikAPI == "" {
-		return
-	}
-	start := p.now()
-	defer func() { observePollDuration("traefik", p.now().Sub(start)) }()
-	statusByURL, err := p.fetchTraefikServerStatus(ctx)
-	if err != nil {
-		debuglog.Debug("frontdesk: poll traefik status", "error", err)
-		p.noteTraefikAPIFailure(ctx)
-		return
-	}
-	members, err := p.store.ListMembers(ctx)
-	if err != nil {
-		return
-	}
-	// Damp the UP->non-UP flip with the same consecutive-miss threshold as health:
-	// Traefik briefly stops listing a member (or marks it DOWN) during a rebuild,
-	// and committing that immediately flaps the badge. "UP" is applied at once
-	// (recovery); a non-UP status is held back until it has been observed
-	// `threshold` polls in a row.
-	threshold := p.healthFailThreshold(ctx)
-	p.mu.Lock()
-	p.traefikAPIFails = 0
-	p.traefikBlanked = false
-	var changed []string
-	for _, m := range members {
-		cur := p.statuses[m.ID]
-		// Key by the same URL BuildTraefikConfig publishes: a legacy row can
-		// still carry userinfo, which the emitted config strips.
-		next := statusByURL[stripUserinfo(m.URL)] // "" when Traefik does not list it
-		if next == "UP" {
-			delete(p.traefikNonUp, m.ID)
-		} else {
-			p.traefikNonUp[m.ID]++
-			if p.traefikNonUp[m.ID] < threshold {
-				continue // tolerate a rebuild blink; keep the last reported status
-			}
-		}
-		if cur.TraefikStatus != next {
-			cur.TraefikStatus = next
-			p.statuses[m.ID] = cur
-			changed = append(changed, m.ID)
-		}
-	}
-	p.mu.Unlock()
-	// Traefik's view caught up to a new/changed member (it needs to re-poll the
-	// config before it lists one), so refresh the UI without a manual reload.
-	for _, id := range changed {
-		p.publishMemberStatus(id)
-	}
-}
-
-// noteTraefikAPIFailure counts consecutive failed polls of Traefik's own API
-// and, once they cross the health-fail threshold, blanks every member's
-// TraefikStatus. Without this the badges freeze at their last value while the
-// API is down, painting a live-looking status from a dead data source; blank
-// renders as the existing faint "unknown". Damped by the same threshold as the
-// UP->DOWN flip so a restart blip of Traefik does not blank the column.
-func (p *Poller) noteTraefikAPIFailure(ctx context.Context) {
-	threshold := p.healthFailThreshold(ctx)
-	p.mu.Lock()
-	if p.traefikBlanked {
-		// Already blanked by an earlier poll in this outage: nothing left to blank,
-		// so stop advancing the counter. A successful poll (which resets both
-		// flags) re-arms this.
-		p.mu.Unlock()
-		return
-	}
-	p.traefikAPIFails++
-	if p.traefikAPIFails < threshold {
-		p.mu.Unlock()
-		return
-	}
-	// Crossed the threshold: blank every live badge. Only members that were
-	// actually polled hold a non-empty TraefikStatus, so iterate the in-memory
-	// statuses directly rather than re-reading the store (a member absent from
-	// this map already renders as "unknown"). Publishing happens after unlock.
-	var changed []string
-	for id, cur := range p.statuses {
-		if cur.TraefikStatus != "" {
-			cur.TraefikStatus = ""
-			p.statuses[id] = cur
-			changed = append(changed, id)
-		}
-		delete(p.traefikNonUp, id)
-	}
-	p.traefikBlanked = true
-	p.mu.Unlock()
-	for _, id := range changed {
-		p.publishMemberStatus(id)
-	}
-}
-
-func (p *Poller) fetchTraefikServerStatus(ctx context.Context) (map[string]string, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, p.traefikAPI+traefikServicesAPI, http.NoBody)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := p.client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("traefik api returned %d", resp.StatusCode)
-	}
-	body, err := io.ReadAll(io.LimitReader(resp.Body, 1<<20))
-	if err != nil {
-		return nil, err
-	}
-	return parseTraefikServerStatus(body)
-}
-
-// traefikServiceInfo is the subset of a Traefik API service object we read.
-type traefikServiceInfo struct {
-	Name         string            `json:"name"`
-	ServerStatus map[string]string `json:"serverStatus"`
-}
-
-// parseTraefikServerStatus extracts the server-URL -> status map for the hotel
-// service from a Traefik /api/http/services response.
-func parseTraefikServerStatus(body []byte) (map[string]string, error) {
-	var services []traefikServiceInfo
-	if err := json.Unmarshal(body, &services); err != nil {
-		// Don't wrap the decoder error: it can echo a fragment of the response.
-		return nil, errors.New("frontdesk: parse traefik services response")
-	}
-	for _, svc := range services {
-		// Traefik names HTTP-provider services like "hotel@http".
-		if svc.Name == traefikServiceName || strings.HasPrefix(svc.Name, traefikServiceName+"@") {
-			return svc.ServerStatus, nil
-		}
-	}
-	return map[string]string{}, nil
-}
-
-// PollVersionsOnce fetches the running version of each member that has a stored
-// admin token, so the UI can flag version mismatches across the group.
-func (p *Poller) PollVersionsOnce(ctx context.Context) {
-	members, err := p.store.ListMembers(ctx)
-	if err != nil {
-		return
-	}
-	for _, m := range members {
-		if !m.HasToken {
-			continue
-		}
-		token, ok, err := p.store.MemberToken(ctx, m.ID)
-		if err != nil || !ok {
-			continue
-		}
-		build, err := p.fetchMemberBuild(ctx, m.URL, token)
-		if err != nil {
-			p.noteVersionFetchFailure(ctx, m, err)
-			// A build we can no longer read is unknown, and the config-sync
-			// gates treat unknown as skewed (fail closed). Keeping the last good
-			// value would let a sync proceed on stale data while the member is
-			// mid-upgrade, which is exactly the window the gate exists for. The
-			// commit is cleared with the version: a commit kept beside a blank
-			// version would outlive the read that vouched for it.
-			p.mu.Lock()
-			cur := p.statuses[m.ID]
-			hadVersion := cur.Version != ""
-			cur.Version = ""
-			cur.Commit = ""
-			p.statuses[m.ID] = cur
-			p.mu.Unlock()
-			if hadVersion {
-				p.publishMemberStatus(m.ID)
-			}
-			continue
-		}
-		p.mu.Lock()
-		cur := p.statuses[m.ID]
-		versionChanged := cur.Version != build.Version || cur.Commit != build.Commit
-		cur.Version = build.Version
-		cur.Commit = build.Commit
-		p.statuses[m.ID] = cur
-		wasAlerting := p.versionFailures[m.ID] >= versionFetchFailThreshold
-		delete(p.versionFailures, m.ID)
-		p.mu.Unlock()
-		if versionChanged {
-			// First successful read (or a build change) for this member: refresh
-			// the UI so the Version column populates without a manual reload. A
-			// commit-only move counts: on a "dev" fleet it is the whole of what
-			// a rebuild changes.
-			p.publishMemberStatus(m.ID)
-		}
-		if wasAlerting {
-			p.recordEvent(ctx, Event{
-				Type:     "version.fetch_recovered",
-				Severity: "success",
-				Source:   "frontdesk-poller",
-				Message:  fmt.Sprintf("Recovered version reads from %s", m.Name),
-				MemberID: m.ID,
-			})
-		}
-	}
-}
-
-// noteVersionFetchFailure tracks consecutive version-fetch failures for a member
-// and raises a single visible warning + event when they cross the threshold. The
-// member's admin token is sent on every attempt, so a persistently failing
-// (possibly hostile or misconfigured) URL is surfaced for the operator rather
-// than retried silently at Debug level forever. The fetch error is logged but
-// never put in the event payload (it can embed a fragment of the member's HTTP
-// response).
-func (p *Poller) noteVersionFetchFailure(ctx context.Context, m *Member, fetchErr error) {
-	p.mu.Lock()
-	p.versionFailures[m.ID]++
-	n := p.versionFailures[m.ID]
-	p.mu.Unlock()
-
-	if n == versionFetchFailThreshold {
-		// Logged separately from the event (which logEvent mirrors) because
-		// the error itself must stay out of the persisted payload.
-		debuglog.Warn("frontdesk: member version fetch failing",
-			"member", m.Name, "consecutive_failures", n, "error", fetchErr)
-		p.recordEvent(ctx, Event{
-			Type:     "version.fetch_failed",
-			Severity: "warning",
-			Source:   "frontdesk-poller",
-			Message:  fmt.Sprintf("Cannot read version from %s after %d attempts; check the member URL", m.Name, n),
-			MemberID: m.ID,
-			Metadata: map[string]any{"consecutive_failures": n},
-		})
-		return
-	}
-	debuglog.Debug("frontdesk: fetch member version", "member", m.Name, "error", fetchErr)
-}
-
-// fetchMemberBuild reads app_version and app_commit from the member's admin
-// settings API. Both ride in one response, so the commit costs no extra call.
-func (p *Poller) fetchMemberBuild(ctx context.Context, baseURL, token string) (memberBuild, error) {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, baseURL+memberSettingsPath, http.NoBody)
-	if err != nil {
-		return memberBuild{}, err
-	}
-	req.Header.Set("Authorization", "Bearer "+token)
-	resp, err := p.client.Do(req)
-	if err != nil {
-		return memberBuild{}, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-	if resp.StatusCode != http.StatusOK {
-		return memberBuild{}, fmt.Errorf("settings api returned %d", resp.StatusCode)
-	}
-	body, err := io.ReadAll(io.LimitReader(resp.Body, 1<<20))
-	if err != nil {
-		return memberBuild{}, err
-	}
-	var payload map[string]any
-	if err := json.Unmarshal(body, &payload); err != nil {
-		// Don't wrap the decoder error: it can echo a fragment of the response.
-		return memberBuild{}, errors.New("frontdesk: parse settings response")
-	}
-	var out memberBuild
-	if v, ok := payload["app_version"].(string); ok {
-		out.Version = v
-	}
-	// A member too old to report app_commit leaves it empty, which buildSkew
-	// reads as "cannot vouch" and falls back to the version comparison.
-	if c, ok := payload["app_commit"].(string); ok {
-		out.Commit = c
-	}
-	return out, nil
-}
-
-// checkConfigStaleness emits a single warning when Traefik has not polled the
-// dynamic config within the configured threshold. It resets on the next poll
-// (RecordConfigPoll), so a recovered provider re-arms the warning.
-func (p *Poller) checkConfigStaleness(ctx context.Context) {
-	threshold := secs(p.settings(ctx).TraefikStaleSecs, 30)
-
-	p.mu.Lock()
-	last := p.lastConfigPollAt
-	notified := p.staleNotified
-	// Never polled yet: arm from "now" so a fresh start does not immediately warn.
-	if last.IsZero() {
-		p.lastConfigPollAt = p.now()
-		p.mu.Unlock()
-		return
-	}
-	stale := p.now().Sub(last) > threshold
-	if stale && !notified {
-		p.staleNotified = true
-	}
-	p.mu.Unlock()
-
-	if stale && !notified {
-		p.recordEvent(ctx, Event{
-			Type: "traefik.stale", Severity: "warning", Source: "frontdesk-poller",
-			Message: fmt.Sprintf("Traefik has not fetched the config for over %s", threshold),
-		})
-	}
-}
-
-// ConfigPollStale reports whether Traefik has stopped fetching the dynamic
-// config past the configured threshold: the same rule checkConfigStaleness
-// alerts on, exposed side-effect-free for the fleet state machine
-// (fleetstate.go). False while unarmed (nothing has ever polled), matching the
-// watchdog's fresh-start grace.
-func (p *Poller) ConfigPollStale(ctx context.Context) bool {
-	threshold := secs(p.settings(ctx).TraefikStaleSecs, 30)
-	p.mu.RLock()
-	last := p.lastConfigPollAt
-	p.mu.RUnlock()
-	return !last.IsZero() && p.now().Sub(last) > threshold
-}
-
-// ConfigPollWarm reports whether the Traefik staleness input has produced a
-// real observation this process: either Traefik has fetched the config at
-// least once, or a full staleness window has elapsed since `since` (process
-// start) without a fetch — at which point the silence is itself the
-// steady-state observation ConfigPollStale will keep reporting. Used by
-// fleetInputsWarm to keep a cold start from reading as a recovery.
-func (p *Poller) ConfigPollWarm(ctx context.Context, since time.Time) bool {
-	p.mu.RLock()
-	armed := !p.lastConfigPollAt.IsZero()
-	p.mu.RUnlock()
-	return armed || p.now().Sub(since) > secs(p.settings(ctx).TraefikStaleSecs, 30)
-}
-
-// checkAutoSyncStale emits a single warning when auto-sync is off and the fleet
-// has not been synced within autoSyncStaleThreshold (see autoSyncStale for the
-// exact rule). Like checkConfigStaleness it de-dups on an in-memory flag so it
-// fires once per stale episode, not every tick; the flag disarms silently when
-// the condition clears (auto-sync re-enabled, or a fresh sync recorded), so a
-// later stale episode alerts again. On a restart the flag resets, so a fleet that
-// is already stale re-alerts once — the same best-effort trade-off as the health
-// and Traefik-staleness checks.
-func (p *Poller) checkAutoSyncStale(ctx context.Context) {
-	cfg, err := p.store.GetAutoSync(ctx)
-	if err != nil {
-		debuglog.Warn("frontdesk: auto-sync staleness: read config", "error", err)
-		return
-	}
-	state, found, err := p.store.GetFleetSyncState(ctx)
-	if err != nil {
-		debuglog.Warn("frontdesk: auto-sync staleness: read fleet sync state", "error", err)
-		return
-	}
-	stale := autoSyncStale(cfg, state.LastRunAt, found, p.now())
-
-	p.mu.Lock()
-	notified := p.autoSyncStaleNotified
-	if stale != notified {
-		p.autoSyncStaleNotified = stale
-	}
-	p.mu.Unlock()
-
-	if stale && !notified {
-		p.recordEvent(ctx, Event{
-			Type: "config.autosync_stale", Severity: "warning", Source: "frontdesk-poller",
-			Message: "Auto-sync is off and the fleet has not been synced in over a day; replicas may be drifting from the primary",
 		})
 	}
 }
